add upgrade to v3 sql statements

+++ lib/db/sql_part.dart
@@ -4,13 +4,16 @@ part of 'database.dart';
 
 extension DatabaseExtensionSQL on DatabaseHelper {
   _onUpgrade(Database db, int oldVersion, int newVersion) async {
-    if (newVersion >= 2) {
+    if (newVersion >= 2 && oldVersion < 2) {
       if (kDebugMode) {
         print('deleting savings table if existent');
       }
       await db.execute('''
           DROP TABLE IF EXISTS saving;
           ''');
+    }
+    if (newVersion >= 3 && oldVersion < 3) {
+      await upgradeToV3(db);
     }
   }
 
@@ -27,7 +30,7 @@ extension DatabaseExtensionSQL on DatabaseHelper {
         color INTEGER NOT NULL,
         balance REAL NOT NULL,
         description TEXT,
-        archived INTEGER NOT NULL,
+        archived INTEGER NOT NULL DEFAULT 0,
         PRIMARY KEY(id),
         CHECK(archived IN (0, 1))
         );
@@ -39,7 +42,7 @@ extension DatabaseExtensionSQL on DatabaseHelper {
         icon INTEGER NOT NULL,
         color INTEGER NOT NULL,
         description TEXT,
-        archived INTEGER NOT NULL,
+        archived INTEGER NOT NULL DEFAULT 0,
         PRIMARY KEY(id),
         CHECK(archived IN (0, 1))
       );
@@ -68,6 +71,7 @@ extension DatabaseExtensionSQL on DatabaseHelper {
         color INTEGER NOT NULL,
         max_value REAL NOT NULL,
         interval_unit TEXT NOT NULL,
+        interval_index INTEGER NOT NULL,
         interval_repetitions INTEGER,
         start_date INTEGER NOT NULL,
         end_date INTEGER,
@@ -125,5 +129,202 @@ extension DatabaseExtensionSQL on DatabaseHelper {
     await db.execute('''
           DROP TABLE IF EXISTS category;
           ''');
+  }
+}
+
+Future<void> upgradeToV3(Database db) async {
+  if (kDebugMode) {
+    print('upgrading to DB V3...');
+  }
+
+  await db.transaction((txn) async {
+    await txn.execute('PRAGMA foreign_keys = OFF;');
+
+    // account
+    await txn.execute('''
+      CREATE TABLE IF NOT EXISTS NEW_account(
+        id INTEGER NOT NULL,
+        name TEXT NOT NULL,
+        icon INTEGER NOT NULL,
+        color INTEGER NOT NULL,
+        balance REAL NOT NULL,
+        description TEXT,
+        archived INTEGER NOT NULL DEFAULT 0,
+        PRIMARY KEY(id),
+        CHECK(archived IN (0, 1))
+        );
+    ''');
+    await txn.execute('''
+      INSERT INTO NEW_account (id, name, icon, color, balance, description) 
+      SELECT id, name, icon, color, balance, description FROM account;
+    ''');
+
+    await txn.execute('''
+      UPDATE NEW_account SET description = NULL
+      WHERE description = '';
+    ''');
+
+    await txn.execute('''
+      DROP TABLE account;
+    ''');
+    await txn.execute('''
+      ALTER TABLE NEW_account RENAME account;
+    ''');
+
+    // category
+    await txn.execute('''
+      CREATE TABLE IF NOT EXISTS NEW_category(
+        id INTEGER NOT NULL,
+        name TEXT NOT NULL,
+        icon INTEGER NOT NULL,
+        color INTEGER NOT NULL,
+        description TEXT,
+        archived INTEGER NOT NULL DEFAULT 0,
+        PRIMARY KEY(id),
+        CHECK(archived IN (0, 1))
+      );
+    ''');
+    await txn.execute('''
+      INSERT INTO NEW_category (id, name, icon, color, balance, description) 
+      SELECT id, name, icon, color, balance, description FROM category;
+    ''');
+    await txn.execute('''
+      UPDATE NEW_category SET description = NULL
+      WHERE description = '';
+    ''');
+    await txn.execute('''
+      DROP TABLE category;
+    ''');
+    await txn.execute('''
+      ALTER TABLE NEW_category RENAME category;
+    ''');
+
+    // singleTransaction
+    await txn.execute('''
+      CREATE TABLE IF NOT EXISTS NEW_singleTransaction(
+        id INTEGER NOT NULL,
+        title TEXT NOT NULL,
+        value REAL NOT NULL,
+        description TEXT,
+        category_id INTEGER NOT NULL,
+        date INTEGER NOT NULL,
+        account1 INTEGER NOT NULL,
+        account2 INTEGER,
+        PRIMARY KEY(id),
+        FOREIGN KEY(category_id) REFERENCES category ON DELETE CASCADE,
+        FOREIGN KEY(account1) REFERENCES account ON DELETE CASCADE,
+        FOREIGN KEY(account2) REFERENCES account ON DELETE CASCADE
+      );
+    ''');
+
+    final List<Map<String, dynamic>> mapSingle = await txn.rawQuery(
+      '''SELECT DISTINCT *, category.icon AS category_icon, category.color AS category_color, category.id AS category_id, category.name AS category_name,
+      account.icon AS account_icon, account.color AS account_color, account.id AS account_id, account.name AS account_name, account.balance AS account_balance,
+      singleTransaction.description AS description, singleTransaction.id AS id
+      FROM singleTransaction, singleTransactionToAccount, category, account
+      WHERE account.id = singleTransactionToAccount.account1_id 
+      AND singleTransactionToAccount.transaction_id = singleTransaction.id
+      AND category.id = singleTransaction.category_id;
+      ''',
+    );
+    for (Map<String, dynamic> transaction in mapSingle) {
+      transaction['date'] =
+          DateTime.parse(transaction['date']).millisecondsSinceEpoch;
+      if (transaction['description'] == '') {
+        String values =
+            '${transaction['id']}, ${transaction['title']}, ${transaction['value']}, ${transaction['category_id']}, ${transaction['date']}, ${transaction['account1_id']}, ${transaction['account2_id']}';
+        await txn.execute('''
+        INSERT INTO NEW_singleTransaction (id, title, value, category_id, date, account1, account2) 
+        VALUES ($values);
+      ''');
+      } else {
+        String values =
+            '${transaction['id']}, ${transaction['title']}, ${transaction['value']}, ${transaction['description']}, ${transaction['category_id']}, ${transaction['date']}, ${transaction['account1_id']}, ${transaction['account2_id']}';
+        await txn.execute('''
+        INSERT INTO NEW_singleTransaction (id, title, value, description, category_id, date, account1, account2) 
+        VALUES ($values);
+      ''');
+      }
+    }
+    await txn.execute('''
+      INSERT INTO NEW_singleTransaction (id, title, value, description, category_id, date, account1, account2) 
+      SELECT id, name, icon, color, balance, description FROM singleTransaction;
+    ''');
+    await txn.execute('''
+      DROP TABLE singleTransaction;
+    ''');
+    await txn.execute('''
+      ALTER TABLE NEW_singleTransaction RENAME singleTransaction;
+    ''');
+
+    // budgets
+    await txn.execute('''
+      CREATE TABLE IF NOT EXISTS NEW_budget(
+        id INTEGER NOT NULL,
+        name TEXT NOT NULL,
+        icon INTEGER NOT NULL,
+        color INTEGER NOT NULL,
+        max_value REAL NOT NULL,
+        interval_unit TEXT NOT NULL,
+        interval_index INTEGER NOT NULL,
+        interval_repetitions INTEGER,
+        start_date INTEGER NOT NULL,
+        end_date INTEGER,
+        description TEXT,
+        PRIMARY KEY(id),
+        CHECK(interval_unit IN ('IntervalUnit.day', 'IntervalUnit.week', 'IntervalUnit.month', 'IntervalUnit.quarter', 'IntervalUnit.year'))
+      );
+    ''');
+
+    final List<Map<String, dynamic>> mapBudgets = await txn.rawQuery(
+      '''SELECT id, name, icon, color, limitXX, interval_unit, interval_repititions, start_date, end_date, description FROM category;
+      ''',
+    );
+    for (Map<String, dynamic> item in mapBudgets) {
+      item['start_date'] =
+          DateTime.parse(item['start_date']).millisecondsSinceEpoch;
+      item['end_date'] =
+          DateTime.parse(item['end_date']).millisecondsSinceEpoch;
+      String values =
+          '${item['id']}, ${item['name']}, ${item['icon']}, ${item['color']}, ${item['limitXX']}, ${item['interval_unit']}, 0, ${item['interval_repititions']}, ${item['start_date']}, ${item['end_date']}, ${item['description'] == '' ? null : item['description']}';
+      await txn.execute('''
+        INSERT INTO categoryBridge (id, name, icon, color, max_value, interval_unit, interval_index, interval_repetitions, start_date, end_date, description) 
+        VALUES ($values);
+      ''');
+    }
+
+    // categoryBridge
+    await txn.execute('''
+      CREATE TABLE IF NOT EXISTS categoryBridge(
+        ancestor_id INTEGER NOT NULL,
+        descendent_id INTEGER NOT NULL,
+        distance INTEGER NOT NULL,
+        PRIMARY KEY(ancestor_id, descendent_id),
+        FOREIGN KEY(ancestor_id) REFERENCES category ON DELETE CASCADE,
+        FOREIGN KEY(descendent_id) REFERENCES category ON DELETE CASCADE
+        );
+    ''');
+    final List<Map<String, dynamic>> mapBridge = await txn.rawQuery(
+      '''SELECT id FROM category;
+      ''',
+    );
+    for (Map<String, dynamic> item in mapBridge) {
+      String values = '${item['id']}, ${item['id']}, 0';
+      await txn.execute('''
+        INSERT INTO categoryBridge (ancestor_id, descendent_id, distance) 
+        VALUES ($values);
+      ''');
+    }
+
+    // final
+    await txn.execute('DROP TABLE categoryToGroup;');
+    await txn.execute('DROP TABLE singleTransactionToAccount;');
+
+    await txn.execute('PRAGMA schema.foreign_key_check;');
+    await txn.execute('PRAGMA foreign_keys = ON;');
+  });
+
+  if (kDebugMode) {
+    print('upgrade to DB V3 done!');
   }
 }
